Added pooling to single header.
